requirements.txt
pandas
jupyter
numpy
mplfinance
requests
pyjwt